Changed to EOS-Payment-Terminal -Removed Private key option.
