+++ EOS-Payment-Terminal/.babelrc
@@ -1,0 +1,7 @@
+{
+  "presets": [
+    "react",
+    "stage-2"
+  ],
+  "plugins": ["emotion"]
+}
