+++ EOS-Payment-Terminal/.npmignore
@@ -1,0 +1,6 @@
+src
+examples
+.babelrc
+.gitignore
+webpack.config.js
+encryptDecrypt.js
