+++ EOS-Payment-Terminal/config.json
@@ -1,0 +1,5 @@
+{
+    "apiKey": "FQK0SYR-W4H4NP2-HXZ2PKH-3J8797N",
+    "saltKey": "bf32eb1e0b28d4b75bb1da9eaa4c5b02"
+}
+
