+++ EOS-Payment-Terminal/package.json
@@ -1,0 +1,65 @@
+{
+  "name": "thin-wallet",
+  "version": "1.0.0",
+  "description": "",
+  "main": "dist/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --mode production",
+    "start": "webpack; webpack-dev-server --color --open --watch --hot --inline && node wsServer/server.js",
+    "transpile": "babel src -d dist --copy-files",
+    "prepublishOnly": "npm run transpile"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "axios": "^0.18.0",
+    "crypto": "^1.0.1",
+    "crypto-js": "^3.1.9-1",
+    "dotenv": "^6.0.0",
+    "eosjs": "^16.0.9",
+    "eosjs-api": "^7.0.4",
+    "eosjs-ecc": "^4.0.4",
+    "express": "^4.16.3",
+    "fetch": "^1.1.0",
+    "file-loader": "^2.0.0",
+    "fs": "0.0.1-security",
+    "https": "^1.0.0",
+    "lodash": "^4.17.11",
+    "node-sass": "^4.9.3",
+    "path": "^0.12.7",
+    "react": "^16.5.1",
+    "react-block-ui": "^1.1.3",
+    "react-bootstrap": "^0.32.4",
+    "react-dom": "^16.5.1",
+    "react-redux": "^5.0.7",
+    "react-spinners": "^0.4.5",
+    "redux": "^4.0.0",
+    "request": "^2.88.0",
+    "sass": "^1.13.4",
+    "sass-loader": "^7.1.0",
+    "scatter-js": "^2.5.2",
+    "socket.io": "^2.1.1",
+    "socket.io-client": "^2.1.1",
+    "style-loader": "^0.23.0",
+    "write-json-file": "^3.0.2"
+  },
+  "devDependencies": {
+    "babel": "^6.23.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "css-loader": "^1.0.0",
+    "dotenv-webpack": "^1.5.7",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
+    "nodemon": "^1.18.4",
+    "redux-devtools-extension": "^2.13.5",
+    "webpack": "^4.19.1",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.8",
+    "webpack-obfuscator": "^0.18.0"
+  }
+}
