+++ EOS-Payment-Terminal/src/.env.api
@@ -1,0 +1,11 @@
+
+PORT=443
+HOST=proxy.eosnode.tools
+SOCKET=http://api.byzanti.ne:9090/
+WS_SOCKET_CHANNEL_TRANSAC_STATUS=isIrreversible
+CHANNEL=user
+TRANSFER=http://api.byzanti.ne:8902/transfer/
+TOKENS=http://api.byzanti.ne:8902/tokensByAccount/
+EOS_PRICE=https://min-api.cryptocompare.com/data/price?fsym=EOS&tsyms=USD,EUR
+TOKENS_PRICE=https://cors-anywhere.herokuapp.com/https://api.newdex.io/v1/ticker/price
+USD_EUR=http://free.currencyconverterapi.com/api/v5/convert?q=EUR_USD&compact=y
