--- stripe-wallet/.babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-    "react",
-    "stage-2"
-  ],
-  "plugins": ["emotion"]
-}
