--- stripe-wallet/.npmignore
@@ -1,6 +1,0 @@
-src
-examples
-.babelrc
-.gitignore
-webpack.config.js
-encryptDecrypt.js
