--- stripe-wallet/server/nonce.json
@@ -1,1 +1,0 @@
-{"nonce":166}
