--- stripe-wallet/src/.env
@@ -1,10 +1,0 @@
-PORT=7777
-HOST=mainnet.libertyblock.io
-SOCKET=http://api.byzanti.ne:8900
-CHANNEL=user
-PROCESS=0000000000000000
-TRANSFER=http://api.byzanti.ne:8902/transfer/
-TOKENS=http://api.byzanti.ne:8902/tokensByAccount/
-EOS_PRICE=https://min-api.cryptocompare.com/data/price?fsym=EOS&tsyms=USD,EUR
-TOKENS_PRICE=https://cors-anywhere.herokuapp.com/https://api.newdex.io/v1/ticker/price
-USD_EUR=http://free.currencyconverterapi.com/api/v5/convert?q=EUR_USD&compact=y
