--- stripe-wallet/src/actions/actionsType.js
@@ -1,3 +1,0 @@
-export const UPDATE_NONCE = 'UPDATE_NONCE';
-export const UPDATE_STATE = 'UPDATE_STATE';
-export const UPDATE_SCATTER = 'UPDATE_SCATTER';
