--- stripe-wallet/src/listing.json
@@ -1,1 +1,0 @@
-{"code":200,"data":[{"symbol":"tea_eos","contract":"linzongsheng","currency":"TEA","price_precision":7,"currency_precision":4},{"symbol":"wizbox_eos","contract":"wizboxairdro","currency":"WIZBOX","price_precision":4,"currency_precision":0},{"symbol":"luck_eos","contract":"eosluckchain","currency":"LUCK","price_precision":6,"currency_precision":4},{"symbol":"epra_eos","contract":"epraofficial","currency":"EPRA","price_precision":6,"currency_precision":4},{"symbol":"dab_eos","contract":"eoscafekorea","currency":"DAB","price_precision":6,"currency_precision":4},{"symbol":"evr_eos","contract":"eosvrtokenss","currency":"EVR","price_precision":6,"currency_precision":4},{"symbol":"deos_eos","contract":"thedeosgames","currency":"DEOS","price_precision":6,"currency_precision":4},{"symbol":"pgl_eos","contract":"prospectorsg","currency":"PGL","price_precision":4,"currency_precision":4},{"symbol":"wecash_eos","contract":"weosservices","currency":"WECASH","price_precision":7,"currency_precision":4},{"symbol":"itecoin_eos","contract":"itecointoken","currency":"ITECOIN","price_precision":6,"currency_precision":4},{"symbol":"trybe_eos","contract":"trybenetwork","currency":"TRYBE","price_precision":5,"currency_precision":4},{"symbol":"faid_eos","contract":"eosfaidchain","currency":"FAID","price_precision":6,"currency_precision":4},{"symbol":"lucky_eos","contract":"eoslucktoken","currency":"LUCKY","price_precision":4,"currency_precision":4},{"symbol":"more_eos","contract":"eosiomoreone","currency":"MORE","price_precision":5,"currency_precision":4},{"symbol":"tpt_eos","contract":"tokendapppub","currency":"TPT","price_precision":6,"currency_precision":4},{"symbol":"pub_eos","contract":"publytoken11","currency":"PUB","price_precision":6,"currency_precision":4},{"symbol":"btn_eos","contract":"eosbuttonbtn","currency":"BTN","price_precision":7,"currency_precision":4},{"symbol":"shape_eos","contract":"gmaslaunches","currency":"SHAPE","price_precision":6,"currency_precision":4},{"symbol":"cat_eos","contract":"tokenbyeocat","currency":"CAT","price_precision":6,"currency_precision":4},{"symbol":"toe_eos","contract":"talkoneos123","currency":"TOE","price_precision":6,"currency_precision":4},{"symbol":"wizz_eos","contract":"wizznetwork1","currency":"WIZZ","price_precision":7,"currency_precision":4},{"symbol":"cet_eos","contract":"eosiochaince","currency":"CET","price_precision":5,"currency_precision":4},{"symbol":"horus_eos","contract":"horustokenio","currency":"HORUS","price_precision":5,"currency_precision":4},{"symbol":"karma_eos","contract":"therealkarma","currency":"KARMA","price_precision":6,"currency_precision":4},{"symbol":"chl_eos","contract":"challengedac","currency":"CHL","price_precision":7,"currency_precision":4},{"symbol":"black_eos","contract":"eosblackteam","currency":"BLACK","price_precision":5,"currency_precision":4},{"symbol":"eosdac_eos","contract":"eosdactokens","currency":"EOSDAC","price_precision":5,"currency_precision":4},{"symbol":"edna_eos","contract":"ednazztokens","currency":"EDNA","price_precision":6,"currency_precision":4},{"symbol":"atd_eos","contract":"eosatidiumio","currency":"ATD","price_precision":6,"currency_precision":4},{"symbol":"meetone_eos","contract":"eosiomeetone","currency":"MEETONE","price_precision":6,"currency_precision":4},{"symbol":"iq_eos","contract":"everipediaiq","currency":"IQ","price_precision":5,"currency_precision":3},{"symbol":"oct_eos","contract":"octtothemoon","currency":"OCT","price_precision":4,"currency_precision":4},{"symbol":"add_eos","contract":"eosadddddddd","currency":"ADD","price_precision":6,"currency_precision":4},{"symbol":"boid_eos","contract":"boidcomtoken","currency":"BOID","price_precision":7,"currency_precision":4},{"symbol":"eox_eos","contract":"eoxeoxeoxeox","currency":"EOX","price_precision":7,"currency_precision":4},{"symbol":"esb_eos","contract":"esbcointoken","currency":"ESB","price_precision":7,"currency_precision":4},{"symbol":"poor_eos","contract":"poormantoken","currency":"POOR","price_precision":7,"currency_precision":4},{"symbol":"eosnts_eos","contract":"eosninetiess","currency":"EOSNTS","price_precision":6,"currency_precision":4},{"symbol":"can_eos","contract":"eoscancancan","currency":"CAN","price_precision":4,"currency_precision":4}]}
